large refactor, split modules, fix tests
