--- test/.eslintrc.json
@@ -1,14 +1,0 @@
-{
-  "parserOptions": {
-    "ecmaVersion": 8
-  },
-  "env": {
-    "node": true,
-    "mocha": true
-  },
-  "globals": {
-    "expect": false,
-    "loadWebExtension": false,
-    "sinon": false
-  }
-}
