+++ test/contentScripts/welcome.test.js
@@ -1,5 +1,3 @@
-require ("../setup.js");
-
 describe("welcome.js", () => {
   // TODO: on completion of welcome page spec
 });
